fix(npc): Fix shop NPC interactions by using PDC tags
